agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
gitpython==3.1.41
google-generativeai==0.3.2
markdown==3.5.2
